Refactor existance file models by transfer each into new module
