+++ apps/file/models/product_file_model.py
@@ -3,23 +3,16 @@ from django.db import models
 from apps.file.models import AbstractFileModel
 
 
-class SimpleFile(AbstractFileModel):
-    """
-    File that has relation one-to-one which relation describe in target table not here.
-    For example banner-image file which are related to a post in our blog.
-    """
-    pass
-
-
 class ProductFile(AbstractFileModel):
     """
     File that are belonged to product images.
+    To let first image has been upload then process of product creation be done, we can have not any product
+        linked to this model.
     """
-    product = models.ForeignKey('product.Product', on_delete=models.CASCADE, related_name='files')
-
-
-class UserProfileImage(AbstractFileModel):
-    """
-    User's profile pictures. each user can have multiple profile pictures, like Telegram.
-    """
-    user = models.ForeignKey('user.User', on_delete=models.CASCADE, related_name='profiles')
+    product = models.ForeignKey(
+        'product.Product',
+        on_delete=models.CASCADE,
+        null=True,
+        default=None,
+        related_name='files'
+    )
